Refactor: Move models in distinct file
